bot/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiscoutbot"
version = "0.1.0"
description = "AI Model Scout WhatsApp Bot"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.*",
    "uvicorn[standard]==0.29.*",
    "apscheduler==3.10.*",
    "httpx==0.27.*",
    "pyyaml==6.0.*",
    "aiosqlite==0.20.*",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "pytz>=2024",
    "smolagents[toolkit]>=1.24",
    "openai>=1.0",
]
